# refactor: optimize file reading for disk-based cache on Unix