--- a/src/query/storages/common/cache/src/providers/disk_cache/mod.rs
+++ b/src/query/storages/common/cache/src/providers/disk_cache/mod.rs
@@ -15,8 +15,10 @@
 #[allow(clippy::module_inception)]
 mod disk_cache;
 mod disk_cache_key;
-mod disk_cache_lru;
+pub mod disk_cache_lru;
 
 pub use disk_cache::*;
 pub use disk_cache_key::DiskCacheKey;
-pub use disk_cache_lru::*;
+pub use disk_cache_lru::LruDiskCache;
+pub use disk_cache_lru::LruDiskCacheBuilder;
+pub use disk_cache_lru::LruDiskCacheHolder;