I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

Optimize file reading for disk-based cache on Unix.
    
For Unix platforms (`#[cfg(unix)]`):
    - Open files with the `O_NOATIME` flag to avoid updating access time.
    - Leverage known file size to streamline reading.
    - Bypass `stat` and `seek` operations used by `std::fs::File::read_to_end` by using `libc::read`


A preliminary benchmark  on my development machine shows:

   (bench code cloud be found at `src/query/storages/common/cache/benches/read_cache_content.rs`)

~~~
read_cache_content_unix_specific_impl/small_file_4KB
                        time:   [1.5226 µs 1.5289 µs 1.5357 µs]
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
read_cache_content_unix_specific_impl/medium_file_64KB
                        time:   [3.8204 µs 3.9527 µs 4.0921 µs]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
read_cache_content_unix_specific_impl/large_file_1MB
                        time:   [45.703 µs 45.875 µs 46.055 µs]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

read_cache_std_lib_read_to_end/small_file_4KB
                        time:   [2.4796 µs 2.4942 µs 2.5113 µs]
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) high mild
  11 (11.00%) high severe
read_cache_std_lib_read_to_end/medium_file_64KB
                        time:   [5.2897 µs 5.5818 µs 5.8790 µs]
Found 12 outliers among 100 measurements (12.00%)
  6 (6.00%) high mild
  6 (6.00%) high severe
read_cache_std_lib_read_to_end/large_file_1MB
                        time:   [46.652 µs 46.840 µs 47.062 µs]
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
~~~


File Size | Unix specific Impl (μs) | Std Lib (μs) | Unix specific Impl is
-- | -- | -- | --
4KB | 1.52 - 1.54 | 2.48 - 2.51 | ~63% faster
64KB | 3.82 - 4.09 | 5.29 - 5.88 | ~38% faster
1MB | 45.7 - 46.1 | 46.7 - 47.1 | roughly the same 


Note that the test machine has sufficient memory, and during testing, the files read are mostly held in the OS cache. For large files, this primarily measures the performance of copying data from kernel space to user space, for small files, the impact of `O_NOATIME`, `stat`, and `seek` operations is more noticeable.


## Tests

- [x] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [x] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/databendlabs/databend/17638)
<!-- Reviewable:end -->